+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "BigTrip",
-  "version": "1.0.4",
+  "version": "1.0.5",
   "private": true,
   "description": "Личный проект «Большое путешествие» от HTML Academy",
   "scripts": {
